calculator/requirements.txt
numpy
tk
Pillow
ttkthemes  # For ThemedTk
re